Restructured pyarts and documentation
