+++ controlfiles-python/artscomponents/surface/Test_iySurfaceFlatReflectivity.py
@@ -6,7 +6,7 @@
 
 import numpy as np
 from pyarts.workspace import Workspace, arts_agenda
-from pyarts.arts import constant
+from pyarts.arts import constants
 from pyarts import xml
 
 
@@ -185,7 +185,7 @@ if __name__ == "__main__":
     wavelength = np.array([600]) * 1e-9
 
     # Frequency grid
-    f_grid = constant.c / wavelength
+    f_grid = constants.c / wavelength
 
     # Set Complex refractive index similar to water at 600 nm
     surface_reflectivity = np.array([[[1.]]])
