+++ python/doc/source/contact.rst
@@ -1,0 +1,14 @@
+Contact
+=======
+
+Github
+------
+`Submit a bug report <https://github.com/atmtools/arts/issues>`
+
+User email list
+---------------
+`Send us an email with user questions <mailto:arts_users.mi@lists.uni-hamburg.de>`_
+
+Developer email list
+--------------------
+`Send us an email about ARTS development <mailto:arts_dev.mi@lists.uni-hamburg.de>`_
