--- python/doc/source/getting_started.rst
@@ -1,25 +1,0 @@
-.. _getting_started:
-
-Getting started
-===============
-
-This *getting started* guide provides a basic introduction to the usage of the
-**arts** Python package. It is separated into two parts. The first one, which
-describes the interactive ARTS interface, provides an introduction on how to
-control ARTS from Python. The second one provides a list of the Python classes
-which are available for the manipulation of built-in ARTS data types.
-
-
-.. toctree::
-   :maxdepth: 2
-
-   arts_interface
-
-
-
-
-
-
-
-
-
