+++ python/doc/source/index.rst
@@ -3,45 +3,16 @@ pyarts
 
 **pyarts** is the Python interface for the
 `Atmospheric Radiative Transfer Simulator <https://www.radiativetransfer.org>`_
-(ARTS). It provides
+(ARTS). It provides an interactive interface to the ARTS engine for running radiative transfer simulations.
 
--  an interactive interface to the ARTS engine for running radiative transfer
-   simulations,
--  Python implementations of many ARTS built-in types for the manipulation
-   of input data and the evaluation simulation results.
-
-Installation
-------------
-
-Pre-compiled binaries
-^^^^^^^^^^^^^^^^^^^^^
-
-Pre-compiled binaries of the ARTS interface including the ARTS engine for
-performing RT simulations can be installed from the
-`Python Package Index <https://pypi.org/project/pyarts>`_ (PyPI):
-
-.. code-block:: bash
-
-    pip install pyarts
-
-
-Building from source
-^^^^^^^^^^^^^^^^^^^^
-
-If you build ARTS from source, you can find instructions on how to install
-PyARTS in the toplevel ``README.md`` file.
-
-
-Getting started
----------------
-
-A basic introduction on how to use the ARTS interface is
-provided in the :ref:`getting_started` section.
+Documentation index
+-------------------
 
 .. toctree::
    :maxdepth: 2
 
-   getting_started
-   workspace_overview
-   modules
+   installation
+   workspace
+   pyarts
    indices
+   contact
