--- python/doc/source/modules.rst
@@ -1,39 +1,0 @@
-API Reference
-=============
-
-ARTS Interface
---------------------
-
-.. toctree::
-   :maxdepth: 2
-
-   pyarts.workspace
-   pyarts_arts
-
-
-ARTS XML I/O
-------------
-
-.. toctree::
-   :maxdepth: 2
-
-   pyarts_xml
-
-
-ARTS Data Types
----------------
-
-.. toctree::
-   :maxdepth: 2
-
-   pyarts_datatypes
-
-
-ARTS Plotting
--------------
-
-.. toctree::
-   :maxdepth: 2
-
-   pyarts_plots
-
