+++ python/doc/source/pyarts.arts_ext.rst
@@ -1,0 +1,6 @@
+pyarts.arts_ext
+===============
+
+.. automodule:: pyarts.arts_ext
+
+.. currentmodule:: pyarts.arts_ext
