+++ python/doc/source/pyarts.cat.rst
@@ -1,0 +1,6 @@
+pyarts.cat
+==========
+
+.. automodule:: pyarts.cat
+
+.. currentmodule:: pyarts.cat
