+++ python/doc/source/pyarts.common.rst
@@ -1,0 +1,6 @@
+pyarts.common
+=============
+
+.. automodule:: pyarts.common
+
+.. currentmodule:: pyarts.common
