+++ python/doc/source/pyarts.environment.rst
@@ -1,0 +1,6 @@
+pyarts.environment
+==================
+
+.. automodule:: pyarts.environment
+
+.. currentmodule:: pyarts.environment
