+++ python/doc/source/pyarts.hitran.rst
@@ -1,0 +1,6 @@
+pyarts.hitran
+=============
+
+.. automodule:: pyarts.hitran
+
+.. currentmodule:: pyarts.hitran
