+++ python/doc/source/pyarts.parser.rst
@@ -1,0 +1,6 @@
+pyarts.parser
+=============
+
+.. automodule:: pyarts.parser
+
+.. currentmodule:: pyarts.parser
