+++ python/doc/source/pyarts.rst
@@ -1,0 +1,19 @@
+API Reference
+=============
+
+.. toctree::
+   :maxdepth: 2
+
+   pyarts.arts
+   pyarts.arts_ext
+   pyarts.cat
+   pyarts.common
+   pyarts.environment
+   pyarts.hitran
+   pyarts.parser
+   pyarts.plots
+   pyarts.sensor
+   pyarts.utils
+   pyarts.workspace
+   pyarts.xml
+
