+++ python/doc/source/pyarts.sensor.rst
@@ -1,0 +1,15 @@
+.. _pyarts_sensor:
+
+pyarts.sensor
+=============
+
+.. automodule:: pyarts.sensor
+
+.. currentmodule:: pyarts.sensor
+
+.. autosummary::
+   :toctree: stubs
+
+   get_f_backend_rel_width
+   get_f_backend_const_width
+
