+++ python/doc/source/pyarts.utils.rst
@@ -1,0 +1,6 @@
+pyarts.utils
+============
+
+.. automodule:: pyarts.utils
+
+.. currentmodule:: pyarts.utils
