--- python/doc/source/pyarts_datatypes.rst
@@ -1,55 +1,0 @@
-.. _pyarts_datatypes:
-
-pyarts.catalogues
-=================
-
-.. automodule:: pyarts.catalogues
-
-.. currentmodule:: pyarts.catalogues
-
-.. autosummary::
-   :toctree: stubs
-
-   Sparse
-
-pyarts.covariancematrix
-=======================
-
-.. automodule:: pyarts.covariancematrix
-
-.. currentmodule:: pyarts.covariancematrix
-
-.. autosummary::
-   :toctree: stubs
-
-   Block
-   CovarianceMatrix
-   plot_covariance_matrix
-
-pyarts.scattering
-=================
-
-.. automodule:: pyarts.scattering
-
-.. currentmodule:: pyarts.scattering
-
-.. autosummary::
-   :toctree: stubs
-
-   SingleScatteringData
-   SpectralSingleScatteringData
-   ScatteringMetaData
-
-pyarts.sensor
-=============
-
-.. automodule:: pyarts.sensor
-
-.. currentmodule:: pyarts.sensor
-
-.. autosummary::
-   :toctree: stubs
-
-   get_f_backend_rel_width
-   get_f_backend_const_width
-
