--- python/doc/source/pyarts_plots.rst
@@ -1,25 +1,0 @@
-pyarts.plots
-============
-
-.. automodule:: pyarts.plots
-
-.. currentmodule:: pyarts.plots
-
-
-Main Functions
---------------
-
-.. autosummary::
-   :toctree: stubs
-
-   plot_arts_lookup
-   polar_ppath
-   polar_ppath_list
-
-Full Modules
-------------
-
-.. autosummary::
-   :toctree: stubs
-
-   pyarts.plots.ppath
