+++ python/doc/source/workspace.rst
@@ -1,0 +1,22 @@
+Workspace Overview
+==================
+
+.. toctree::
+    :hidden:
+
+    workspace.groups
+    workspace.methods
+    workspace.variables
+    workspace.agendas
+
+.. rubric:: Workspace Groups
+.. include:: workspace.groups.auto.rst
+
+.. rubric:: Workspace Variables
+.. include:: workspace.variables.auto.rst
+
+.. rubric:: Workspace Methods
+.. include:: workspace.methods.auto.rst
+
+.. rubric:: Workspace Agendas
+.. include:: workspace.agendas.auto.rst
