--- python/doc/source/workspace_overview.rst
@@ -1,22 +1,0 @@
-Workspace Overview
-==================
-
-.. toctree::
-    :hidden:
-
-    workspace_groups_overview
-    workspace_methods_overview
-    workspace_variables_overview
-    workspace_agendas_overview
-
-.. rubric:: Workspace Groups
-.. include:: auto_wsg_overview.rst
-
-.. rubric:: Workspace Methods
-.. include:: auto_wsm_overview.rst
-
-.. rubric:: Workspace Variables
-.. include:: auto_wsv_overview.rst
-
-.. rubric:: Workspace Agendas
-.. include:: auto_wsa_overview.rst
