--- python/pyarts/catalogues.py
@@ -1,115 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Implementation of classes to handle various catalogue information.
-
-"""
-
-
-import numpy as np
-import scipy.sparse
-
-__all__ = ['Sparse',
-           ]
-
-
-class Sparse(scipy.sparse.csc_matrix):
-    """Wrapper around :class:`scipy.sparse.csc_matrix`.
-
-    This class wraps around the SciPy Compressed Sparse Column matrix. The
-    usage is exactly the same, but support for reading and writing XML files
-    is added. Also additional attributes were added to map the ARTS
-    implementation of :arts:`Sparse`.
-
-    """
-    @property
-    def nrows(self):
-        """Number of rows."""
-        return self.shape[0]
-
-    @property
-    def ncols(self):
-        """Number of columns."""
-        return self.shape[0]
-
-    @property
-    def rowindex(self):
-        """Row indices to locate data in matrix."""
-        return self.tocoo().row
-
-    @property
-    def colindex(self):
-        """Column indices to locate data in matrix."""
-        return self.tocoo().col
-
-    @property
-    def sparsedata(self):
-        """Data value at specified positions in matrix."""
-        return self.tocoo().data
-
-    @classmethod
-    def from_xml(cls, xmlelement):
-        """Loads a Sparse object from an existing file."""
-
-        binaryfp = xmlelement.binaryfp
-        nelem = int(xmlelement[0].attrib['nelem'])
-        nrows = int(xmlelement.attrib['nrows'])
-        ncols = int(xmlelement.attrib['ncols'])
-
-        if binaryfp is None:
-            rowindex = np.fromstring(xmlelement[0].text, sep=' ').astype(int)
-            colindex = np.fromstring(xmlelement[1].text, sep=' ').astype(int)
-            sparsedata = np.fromstring(xmlelement[2].text, sep=' ')
-        else:
-            rowindex = np.fromfile(binaryfp, dtype='<i4', count=nelem)
-            colindex = np.fromfile(binaryfp, dtype='<i4', count=nelem)
-            sparsedata = np.fromfile(binaryfp, dtype='<d', count=nelem)
-
-        return cls((sparsedata, (rowindex, colindex)), [nrows, ncols])
-
-    def write_xml(self, xmlwriter, attr=None):
-        """Write a Sparse object to an ARTS XML file."""
-
-        # Get ARTS-style information from CSC matrix.
-        nrows = self.shape[0]
-        ncols = self.shape[1]
-        rowindex = self.tocoo().row
-        colindex = self.tocoo().col
-        sparsedata = self.tocoo().data
-
-        precision = xmlwriter.precision
-
-        if attr is None:
-            attr = {}
-
-        attr['nrows'] = nrows
-        attr['ncols'] = ncols
-
-        xmlwriter.open_tag('Sparse', attr)
-
-        binaryfp = xmlwriter.binaryfilepointer
-
-        if binaryfp is None:
-            xmlwriter.open_tag('RowIndex', {'nelem': rowindex.size})
-            for i in rowindex:
-                xmlwriter.write('%d' % i + '\n')
-            xmlwriter.close_tag()
-            xmlwriter.open_tag('ColIndex', {'nelem': colindex.size})
-            for i in colindex:
-                xmlwriter.write('%d' % i + '\n')
-            xmlwriter.close_tag()
-            xmlwriter.open_tag('SparseData', {'nelem': sparsedata.size})
-            for i in sparsedata:
-                xmlwriter.write(('%' + precision) % i + '\n')
-            xmlwriter.close_tag()
-            xmlwriter.close_tag()
-        else:
-            xmlwriter.open_tag('RowIndex', {'nelem': rowindex.size})
-            np.array(rowindex, dtype='i4').tofile(binaryfp)
-            xmlwriter.close_tag()
-            xmlwriter.open_tag('ColIndex', {'nelem': colindex.size})
-            np.array(colindex, dtype='i4').tofile(binaryfp)
-            xmlwriter.close_tag()
-            xmlwriter.open_tag('SparseData', {'nelem': sparsedata.size})
-            np.array(sparsedata, dtype='d').tofile(binaryfp)
-            xmlwriter.close_tag()
-            xmlwriter.close_tag()
