--- python/pyarts/classes.py
@@ -1,1 +1,0 @@
-raise ImportError("pyarts.classes has been moved to pyarts.arts")
