+++ src/python_interface/py_constant.cpp
@@ -1,23 +1,19 @@
 #include <arts_constants.h>
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
+#include <pybind11/pytypes.h>
 
 namespace Python {
 namespace py = pybind11;
 
 struct ConstantDummy {};
 
-#define PythonInterfaceConstant(name) \
-  constant.def_readonly_static(       \
-      #name, &Constant::name, py::doc(var_string("Value: ", Constant::name).c_str()))
+#define PythonInterfaceConstant(name) constants.attr(#name) = Constant::name;
 
 void py_constants(py::module_& m) {
-  auto constant = py::class_<ConstantDummy>(m, "constant");
-  constant.doc() = R"--(Contains constant internal values of Arts
+  auto constants = m.def_submodule("constants", R"--(Contain copies of constants of Arts internals
+)--");
 
-Note that this "class" simply contains static data and that it should NOT
-be initialized
-)--";
   PythonInterfaceConstant(pi);
   PythonInterfaceConstant(c);
   PythonInterfaceConstant(h);
